fixed the frontend build for autologin
